PLAT-764 - Rename the project to CDFlow Commands
The name infra-deployer doesn't really fit with the release and
destroy actions which this also performs so we're giving it a more
general name.
