+++ requirements.txt
@@ -1,8 +1,3 @@
-# TODO: pin dependency versions
-
 boto3==1.4.0
 credstash==1.12.0
 docopt==0.6.2
-nose==1.3.7
-placebo
-testfixtures
